fixes circular deps
