--- webapp/src/main/webapp/src/app/assessments/model/claim-score.model.ts
@@ -1,10 +1,0 @@
-import { ExamStatisticsLevel } from './exam-statistics.model';
-
-/**
- * @deprecated use ScoreStatistics
- */
-export class ClaimStatistics {
-  id: number;
-  levels: ExamStatisticsLevel[];
-  percents: ExamStatisticsLevel[];
-}
