+++ webapp/src/main/webapp/src/app/assessments/model/exam-statistics.model.ts
@@ -1,5 +1,3 @@
-import { ClaimStatistics } from './claim-score.model';
-
 /**
  * @deprecated use ScoreStatistics
  */
@@ -19,4 +17,13 @@ export class ExamStatisticsLevel {
   id: number;
   value: number;
   suffix: string;
+}
+
+/**
+ * @deprecated use ScoreStatistics
+ */
+export class ClaimStatistics {
+  id: number;
+  levels: ExamStatisticsLevel[];
+  percents: ExamStatisticsLevel[];
 }
