+++ webapp/src/main/webapp/src/app/assessments/results/view/export-results.ts
@@ -1,0 +1,4 @@
+export interface ExportResults {
+  exportToCsv(): void;
+  hasDataToExport(): boolean;
+}
