+++ webapp/src/main/webapp/src/app/assessments/results/view/results-by-item/results-by-item.component.ts
@@ -7,8 +7,8 @@ import { ExamStatisticsCalculator } from '../../exam-statistics-calculator';
 import { Exam } from '../../../model/exam';
 import { Assessment } from '../../../model/assessment';
 import { RequestType } from '../../../../shared/enum/request-type.enum';
-import { ExportResults } from '../../assessment-results.component';
 import { AssessmentExporter } from '../../../assessment-exporter.interface';
+import { ExportResults } from '../export-results';
 
 @Component({
   selector: 'results-by-item',
